Project api sync to project in hrms api

+++ database/migrations/2024_04_12_002657_tbl_overtime_removeemployeeid.php
@@ -13,8 +13,10 @@ return new class extends Migration
     {
 
         Schema::table('overtime', function (Blueprint $table) {
-            $table->dropForeign('overtime_employee_id_foreign');
-            $table->dropColumn('employee_id');
+            if (Schema::hasColumn('overtime', 'employee_id')) {
+                $table->dropForeign('overtime_employee_id_foreign');
+                $table->dropColumn('employee_id');
+            }
         });
     }
 
@@ -24,7 +26,7 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('overtime', function (Blueprint $table) {
-            $table->foreignId('employee_id');
+            $table->foreignId('employee_id')->nullable();
             $table->foreign('employee_id')->references('id')->on('employees');
         });
     }
